+++ CHANGELOG.md
@@ -1,12 +1,12 @@
 ## HEAD (unreleased)
 
 - Fix double output bug (https://github.com/zombocom/dead_end/pull/99)
-- Fix bug causing poor results (fix #95, fix #88) ()
+- Fix bug causing poor results (fix #95, fix #88) (https://github.com/zombocom/dead_end/pull/96)
 - [Breaking] Remove previously deprecated `require "dead_end/fyi"` interface (https://github.com/zombocom/dead_end/pull/94)
 - DeadEnd is now fired on EVERY syntax error (https://github.com/zombocom/dead_end/pull/94)
 - Output format changes:
   - Parse errors emitted per-block rather than for the whole document (https://github.com/zombocom/dead_end/pull/94)
-  - The "banner" is now based on lexical analysis rather than parser regex (fix #68, fix #87) ()
+  - The "banner" is now based on lexical analysis rather than parser regex (fix #68, fix #87) (https://github.com/zombocom/dead_end/pull/96)
 
 ## 2.0.2